+++ tests/test_flow.ts
@@ -1016,7 +1016,7 @@ describe('flows', () => {
       });
     }
 
-    await flow.add(
+    const { job } = await flow.add(
       {
         name: 'parent-job',
         queueName: parentQueueName,
@@ -1037,6 +1037,16 @@ describe('flows', () => {
     await running;
     await childrenWorker.close();
     await processingParent;
+
+    const { children } = await flow.getFlow({
+      queueName: parentQueueName,
+      id: job.id,
+    });
+
+    for (const { job: child } of children) {
+      expect(child).to.not.be.undefined;
+    }
+
     await parentWorker.close();
     await queueScheduler.close();
     await queueEvents.close();